Configure vlan on host
This commit includes the configuration of vlans on the host
for virtual machines. By configuring a bridge on the br0 interface
we can host the virtual machine vlan and by setting a vlan on the
logical device we can include it into the management vlan.
All variables are included from playbook