roles/cotainer_file_read: port to Python 3
Make it work on Python3, even if deprecated with the new `ssh_lxc`
connection plugin.

Now you can use directly the `slurp` native module with delegation to
the lxc guest.

May be useful only in case where it is impossible to get python3 or
other ansible requirements on the lxc guest.